+++ docs/guides/changelog.md
@@ -9,6 +9,7 @@ description: |-
 
 ## Unreleased
 
+- Add RMON event configuration to `nxos_snmp` resource and data source
 - Add management interface configuration to `nxos_system` resource and data source
 - Add LLDP configuration to `nxos_system` resource and data source
 - Add platform entity configuration to `nxos_system` resource and data source